--- pokedex13/src/pages/Header/Header.js
@@ -1,11 +1,0 @@
-import React from "react";
-import { HeaderContainer } from "./style";
-
-const Header = () => {
-  return (
-    <HeaderContainer />
-  )
-  
-};
-
-export default Header;
