--- pokedex13/src/pages/Header/style.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const HeaderContainer = styled.div`
-    background: url('https://i2.wp.com/multarte.com.br/wp-content/uploads/2019/03/pokemon-png-logo.png?fit=2000%2C736&ssl=1'); 
-    background-repeat: no-repeat;
-    background-color: red;
-    background-size: 300px 100px;
-    background-position: center;
-    display: flex;
-    height: 18vh;
-`;
