--- pokedex13/src/pages/Pokedex/style.js
@@ -1,27 +1,0 @@
-import styled from "styled-components";
-
-export const Title = styled.h1`
-  text-align: center;
-`;
-
-export const ContainerButton = styled.div`
-  display: flex;
-  justify-content: flex-start;
-  flex-wrap: wrap;
-  padding: 10px;
-
-  button {
-    background-color: #ffdb04;
-    color: #2c71b6;
-    font-weight: bold;
-    border-radius: 5px;
-    border: 2.3px solid #2c71b6;
-    padding: 8px;
-
-    &:hover {
-      background-color: #f8f8ff;
-      cursor: pointer;
-      transition: 200ms;
-    }
-  }
-`;
