--- pokedex13/src/parameters/baseURL.js
@@ -1,1 +1,0 @@
-export const baseURL = "https://pokeapi.co/api/v2/pokemon"
