+++ src/components/Header/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { HeaderContainer } from './style';
+
+const Header = () => {
+  return <HeaderContainer />;
+};
+
+export default Header;
