+++ src/components/Header/style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const HeaderContainer = styled.div`
+    background: url('https://o.remove.bg/downloads/f35fcada-0ad0-48fe-9980-637a9b9d627a/9ceg4vlucdeq5vshq8vy-removebg-preview.png'); 
+    background-repeat: no-repeat;
+    background-color: red;
+    background-size: 300px 150px;
+    background-position: center;
+    display: flex;
+    height: 18vh;
+`;
