+++ src/pages/Details/style.js
