+++ src/resources/baseURL.js
@@ -1,0 +1,1 @@
+export const baseURL = 'https://pokeapi.co/api/v2/pokemon';
